@@ -9,34 +9,102 @@
 
   <packaging>jar</packaging>
 
+  <prerequisites>
+    <maven>3.0</maven>
+  </prerequisites>
+
+  <properties>
+    <grizzly.version>2.3.2</grizzly.version>
+    <jersey.version>1.17.1</jersey.version>
+    <jersey-mustache.version>1.0.0</jersey-mustache.version>
+    <jackson.version>2.2.1</jackson.version>
+    <mustache.version>0.8.12</mustache.version>
+    <hibernate.version>4.2.1.Final</hibernate.version>
+    <mysql-connector.version>5.1.25</mysql-connector.version>
+  </properties>
+
   <dependencies>
     <dependency>
       <groupId>com.sun.jersey</groupId>
       <artifactId>jersey-server</artifactId>
-      <version>1.17.1</version>
     </dependency>
     <dependency>
       <groupId>com.sun.jersey</groupId>
       <artifactId>jersey-grizzly2</artifactId>
-      <version>1.17.1</version>
+    </dependency>
+    <dependency>
+      <groupId>org.eluder.jersey</groupId>
+      <artifactId>jersey-mustache</artifactId>
+      <version>${jersey-mustache.version}</version>
     </dependency>
     <dependency>
       <groupId>com.fasterxml.jackson.core</groupId>
       <artifactId>jackson-databind</artifactId>
-      <version>2.1.4</version>
+      <version>${jackson.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.hibernate</groupId>
+      <artifactId>hibernate-core</artifactId>
+      <version>${hibernate.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.hibernate</groupId>
+      <artifactId>hibernate-c3p0</artifactId>
+      <version>${hibernate.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>mysql</groupId>
+      <artifactId>mysql-connector-java</artifactId>
+      <version>${mysql-connector.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>commons-cli</groupId>
+      <artifactId>commons-cli</artifactId>
+      <version>1.2</version>
     </dependency>
   </dependencies>
+  
+  <dependencyManagement>
+    <dependencies>
+      <dependency>
+        <groupId>com.sun.jersey</groupId>
+        <artifactId>jersey-server</artifactId>
+        <version>${jersey.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>com.sun.jersey</groupId>
+        <artifactId>jersey-grizzly2</artifactId>
+        <version>${jersey.version}</version>
+      </dependency>
+      <dependency>
+      	<groupId>org.glassfish.grizzly</groupId>
+      	<artifactId>grizzly-http</artifactId>
+      	<version>${grizzly.version}</version>
+      </dependency>
+      <dependency>
+      	<groupId>org.glassfish.grizzly</groupId>
+      	<artifactId>grizzly-http-server</artifactId>
+      	<version>${grizzly.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>com.github.spullara.mustache.java</groupId>
+        <artifactId>compiler</artifactId>
+        <version>${mustache.version}</version>
+      </dependency>
+    </dependencies>
+  </dependencyManagement>
 
   <build>
+    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <inherited>true</inherited>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
-        <version>2.3.2</version>
+        <version>3.1</version>
         <configuration>
-          <source>1.6</source>
-          <target>1.6</target>
+          <source>1.7</source>
+          <target>1.7</target>
           <optimize>true</optimize>
           <debug>false</debug>
         </configuration>
