@@ -5,9 +5,9 @@ from django.db import models
 class World(models.Model):
   randomnumber = models.IntegerField()
   class Meta:
-    db_table = 'world'
+    db_table = 'World'
 
 class Fortune(models.Model):
-  message = models.CharField()
+  message = models.CharField(max_length=65535)
   class Meta:
-    db_table = 'fortune'
+    db_table = 'Fortune'
