@@ -5,17 +5,20 @@ from django.http import HttpResponse
 from django.core import serializers
 from world.models import World, Fortune
 from django.shortcuts import render
-import ujson
+from ujson import dumps as uj_dumps
 import random
+import sys
 from operator import attrgetter
-import numpy.random as nprnd
 from functools import partial
+
+if sys.version_info[0] == 3:
+  xrange = range
 
 def json(request):
   response = {
     "message": "Hello, World!"
   }
-  return HttpResponse(ujson.dumps(response), mimetype="application/json")
+  return HttpResponse(uj_dumps(response), mimetype="application/json")
 
 def db(request):
   queries = int(request.GET.get('queries', 1))
@@ -31,19 +34,16 @@ def db(request):
   # one can eliminate dereferences by storing the end dereferenced thing in an identifier
   g = World.objects.get
   #r = random.randint
-  # but wait! there's more!
-  #http://stackoverflow.com/questions/4172131/create-random-list-of-integers-in-python
-  #r = nprnd.randint
   # but wait!  there's more!  if we're calling a function over and over with the same parameters, 
   # we can use even more function magic.
-  rp = partial(nprnd.randint, 1, 10000)
+  rp = partial(random.randint, 1, 10000)
   # now we're ready to write our awesome query iterator thingy
   # first of all, we know the id's correspond to the random number we're picking, so we can create
   # dictionaries on the fly instead of serializing later
   # by creating dicts, we don't need to user the model serializer, which is probably slow and only appropriate
   # for complicated serializations of joins and crazy query sets etc
   # test xrange vs range if the query number is gigantic
-  worlds = ujson.dumps([{'id' : r, 'randomNumber' : g(id=r).randomnumber} for r in [rp() for q in xrange(queries)]])  
+  worlds = uj_dumps([{'id' : r, 'randomNumber' : g(id=r).randomnumber} for r in [rp() for q in xrange(queries)]])  
   return HttpResponse(worlds, mimetype="application/json")
 
 def fortunes(request):
@@ -54,3 +54,18 @@ def fortunes(request):
 
   context = {'fortunes': fortunes}
   return render(request, 'fortunes.html', context)
+
+def update(request):
+  queries = int(request.GET.get('queries', 1))
+  g = World.objects.get
+  rp = partial(random.randint, 1, 10000)
+  
+  worlds = []
+  for r in [rp() for q in xrange(queries)]:
+    w = g(id=r)
+    w.randomnumber=rp()
+    w.save()
+
+    worlds.append({'id' : r, 'randomNumber' : w.randomnumber})
+
+  return HttpResponse(worlds, mimetype="application/json")
