@@ -5,11 +5,11 @@
 
 var cluster = require('cluster')
   , numCPUs = require('os').cpus().length
+  , windows = require('os').platform() == 'win32'
   , express = require('express')
   , mongoose = require('mongoose')
   , async = require('async')
   , conn = mongoose.connect('mongodb://localhost/hello_world')
-  , Mapper = require('mapper')
   , connMap = { user: 'benchmarkdbuser', password: 'benchmarkdbpass', database: 'hello_world', host: 'localhost' };
 
 var Schema = mongoose.Schema
@@ -21,9 +21,12 @@ var WorldSchema = new Schema({
 }, { collection : 'world' });
 var MWorld = conn.model('World', WorldSchema);
 
-Mapper.connect(connMap, {verbose: false, strict: false});
-var World = Mapper.map("World", "id", "randomNumber");
-var Fortune = Mapper.map("Fortune", "id", "message");
+if (!windows) {
+  var Mapper = require('mapper');
+  Mapper.connect(connMap, {verbose: false, strict: false});
+  var World = Mapper.map("World", "id", "randomNumber");
+  var Fortune = Mapper.map("Fortune", "id", "message");
+}
 
 if (cluster.isMaster) {
   // Fork workers.
@@ -81,6 +84,8 @@ if (cluster.isMaster) {
   });
 
   app.get('/mysql-orm', function(req, res) {
+    if (windows) return res.send(501, 'Not supported on windows');
+    
     var queries = req.query.queries || 1
       , worlds  = []
       , queryFunctions = [];
@@ -100,6 +105,8 @@ if (cluster.isMaster) {
   });
 
   app.get('/fortune', function(req, res) {
+    if (windows) return res.send(501, 'Not supported on windows');
+    
     Fortune.all(function (err, fortunes) {
       var newFortune = {id: 0, message: "Additional fortune added at request time."};
       fortunes.push(newFortune);
@@ -112,6 +119,72 @@ if (cluster.isMaster) {
   function sortFortunes(a, b) {
     return (a.message < b.message) ? -1 : (a.message > b.message) ? 1 : 0;
   }
+
+  app.get('/mongoose-update', function(req, res) {
+    var queries = req.query.queries || 1
+      , selectFunctions = [];
+
+    queries = Math.min(queries, 500);
+
+    for (var i = 1; i <= queries; i++ ) {
+      selectFunctions.push(function(callback) {
+        MWorld.findOne({ id: Math.floor(Math.random() * 10000) + 1 }).exec(callback);
+      });
+    }
+
+    async.parallel(selectFunctions, function(err, worlds) {
+      var updateFunctions = [];
+
+      for (var i = 0; i < queries; i++) {
+        (function(i){
+          updateFunctions.push(function(callback){
+            worlds[i].randomNumber = Math.ceil(Math.random() * 10000);
+            MWorld.update({
+              id: worlds[i]
+            }, {
+              randomNumber: worlds[i].randomNumber
+            }, callback);
+          });
+        })(i);
+      }
+
+      async.parallel(updateFunctions, function(err, updates) {
+        res.send(worlds);
+      });
+    });
+  });
+
+  app.get('/mysql-orm-update', function(req, res) {
+    if (windows) return res.send(501, 'Not supported on windows');
+
+    var queries = req.query.queries || 1
+      , selectFunctions = [];
+
+    queries = Math.min(queries, 500);
+
+    for (var i = 1; i <= queries; i++ ) {
+      selectFunctions.push(function(callback) {
+        World.findById(Math.floor(Math.random() * 10000) + 1, callback);
+      });
+    }
+
+    async.parallel(selectFunctions, function(err, worlds) {
+      var updateFunctions = [];
+
+      for (var i = 0; i < queries; i++) {
+        (function(i){
+          updateFunctions.push(function(callback){
+            worlds[i].randomNumber = Math.ceil(Math.random() * 10000);
+            World.save(worlds[i], callback);
+          });
+        })(i);
+      }
+
+      async.parallel(updateFunctions, function(err, updates) {
+        res.send(worlds);
+      });
+    });   
+  });
 
   app.listen(8080);
 }
