@@ -1,11 +1,14 @@
 <?php
 
+use Phalcon\Mvc\View,
+    Phalcon\Mvc\Model\Resultset;
+
 class BenchController extends \Phalcon\Mvc\Controller
 {
     public function initialize()
     {
         // views must be renderd explicitly. safes processing time when not needed.
-        $this->view->disable();
+        $this->view->setRenderLevel(View::LEVEL_LAYOUT);
     }
 
     public function jsonAction() {
@@ -15,7 +18,9 @@ class BenchController extends \Phalcon\Mvc\Controller
     }
 
     public function dbAction() {
-        $queries = $this->getQueryOrDefault('queries', 1);
+
+        $queries = $this->request->getQuery('queries', null, 1);
+
         $worlds = array();
 
         for ($i = 0; $i < $queries; ++$i) {
@@ -26,50 +31,59 @@ class BenchController extends \Phalcon\Mvc\Controller
     }
 
     public function fortunesAction() {
+
         // since the resultset is immutable get an array instead
         // so we can add the new fortune
         $fortunes = Fortunes::find()->toArray();
+
         $fortunes[] = array(
             'id' => 0,
             'message' => 'Additional fortune added at request time.'
         );
 
         usort($fortunes, function($left, $right) {
-            if ($left['message'] === $right['message']) {
+            $l = $left['message'];
+            $r = $right['message'];
+            if ($l === $r) {
                 return 0;
-            } else if ($left['message'] > $right['message']) {
-                return 1;
             } else {
-                return -1;
+                if ($l > $r) {
+                    return 1;
+                } else {
+                    return -1;
+                }
             }
         });
 
-        return $this->sendContentAsText(
-            $this->view->getRender('bench', 'fortunes', array(
-                'fortunes' => $fortunes
-            ))
-        );
+        $this->response->setHeader("Content-Type", "text/html; charset=utf-8");
+
+        $this->view->fortunes = $fortunes;
     }
 
-    private function getQueryOrDefault($query, $default) {
-        return $this->request->getQuery($query) !== null
-            ? $this->request->getQuery($query)
-            : $default;
-    }
+    public function updateAction() {
 
-    private function sendContentAsText($content) {
-        $response = new Phalcon\Http\Response();
-        $response->setStatusCode(200, "OK");
-        $response->setHeader("Content-Type", "text/html; charset=utf-8");
-        $response->setContent($content);
-        return $response;
+        $queries = $this->request->getQuery('queries', null, 1);
+        if($queries < 1) {
+            $queries = 1;
+        } else if ($queries > 500) {
+            $queries = 500;
+        }
+
+        $worlds = array();
+
+        for ($i = 0; $i < $queries; ++$i) {
+            $world = Worlds::findFirst(mt_rand(1, 10000));
+            $world->randomNumber = mt_rand(1, 10000);
+            $world->save();
+            $worlds[] = $world;
+        }
+
+        return $this->sendContentAsJson($worlds);
     }
 
     private function sendContentAsJson($content) {
-        $response = new Phalcon\Http\Response();
-        $response->setStatusCode(200, "OK");
+        $response = new Phalcon\Http\Response(json_encode($content));
         $response->setHeader("Content-Type", "application/json");
-        $response->setContent(json_encode($content));
         return $response;
     }
 }
