@@ -1,5 +1,5 @@
 worker_processes 8
-listen "/tmp/.sock", :backlog => 256
+listen "/tmp/.sock", :backlog => 4096
 
 preload_app true
 GC.respond_to?(:copy_on_write_friendly=) and
