@@ -6,10 +6,10 @@ import re
 def start(args):
 
   try:
-    subprocess.check_call("rvm jruby-1.7.3 do bundle install --gemfile=Gemfile-jruby", shell=True, cwd="rack")
+    subprocess.check_call("rvm jruby-1.7.4 do bundle install --gemfile=Gemfile-jruby", shell=True, cwd="rack")
     subprocess.check_call("cp Gemfile-jruby Gemfile", shell=True, cwd="rack")
     subprocess.check_call("cp Gemfile-jruby.lock Gemfile.lock", shell=True, cwd="rack")
-    subprocess.check_call("rvm jruby-1.7.3 do warble war", shell=True, cwd="rack")
+    subprocess.check_call("rvm jruby-1.7.4 do warble war", shell=True, cwd="rack")
     subprocess.check_call("rm -rf $RESIN_HOME/webapps/*", shell=True)
     subprocess.check_call("cp rack.war $RESIN_HOME/webapps/rack.war", shell=True, cwd="rack")
     subprocess.check_call("$RESIN_HOME/bin/resinctl start", shell=True)
